Split integration tests into files per replicator/extractor, moved test specific fixtures out of conftest
